# API-Gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Freelance Marketplace.",
        "title": "GigForge",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "gigforge.jasim.online",
    "basePath": "/",
    "paths": {
...
        "/profile/client": {
            "get": {
                "description": "Get the profile of a client",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Get client profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    }
                }
            }
        },
        "/profile/company": {
            "put": {
                "description": "Update the company details for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update company details",
                "parameters": [
                    {
                        "description": "Company Details",
                        "name": "CompanyDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UpdateCompanyDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    }
                }
            }
        },
        "/profile/company/contact": {
            "put": {
                "description": "Update the company contact details for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update company contact details",
                "parameters": [
                    {
                        "description": "Company Contact Details",
                        "name": "CompanyContact",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UpdateCompanyContact"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    }
                }
            }
        },
        "/profile/description": {
            "put": {
                "description": "Update profile description for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile description",
                "parameters": [
                    {
                        "description": "Profile Details",
                        "name": "Profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.Profile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    }
                }
            },
            "post": {
                "description": "Add profile description for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Add profile description",
                "parameters": [
                    {
                        "description": "Profile Details",
                        "name": "Profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.Profile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    }
                }
            }
        },
        "/profile/education": {
            "post": {
                "description": "Add education details for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Add education details",
                "parameters": [
                    {
                        "description": "Education Details",
                        "name": "Education",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.Education"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    }
                }
            }
        },
        "/profile/education/{id}": {
            "put": {
                "description": "Update education details for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update education details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Education ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Education Details",
                        "name": "Education",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.Education"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete education details for the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Delete education details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Education ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    }
                }
            }
        },
        "/profile/experience": {
            "post": {
                "description": "Add experience for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Add experience",
                "parameters": [
                    {
                        "description": "Experience Details",
                        "name": "Experience",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.Experience"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    }
                }
            }
        },
...
        "/profile/freelancer": {
            "get": {
                "description": "Get the profile of a freelancer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Get freelancer profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    }
                }
            }
        },
        "/profile/notifications": {
            "get": {
                "description": "Get notifications for the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Get user notifications",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "UserId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/res.CommonRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "req.UpdateCompanyContact": {
            "type": "object",
            "properties": {
                "OwnerName": {
                    "type": "string"
                },
                "address": {
                    "$ref": "#/definitions/req.Address"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "req.UpdateCompanyDetails": {
            "type": "object",
            "properties": {
                "NumberOfEmployees": {
                    "type": "integer"
                },
                "company": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "tagline": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "res.CommonRes": {
            "type": "object",
            "properties": {
                "body": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ClientAccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "FreelancerAccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}